Decouple buffering logic from MaskingState for performance's sake
